+++ tla-types/src/test/scala/at/forsyte/apalache/tla/typecomp/TestVariantBuilder.scala
@@ -7,36 +7,12 @@ import org.scalacheck.Gen
 import org.scalatestplus.junit.JUnitRunner
 import scalaz.unused
 
-import scala.collection.immutable.SortedMap
-
 @RunWith(classOf[JUnitRunner])
 class TestVariantBuilder extends BuilderTest {
-
-  implicit val tagGen: Gen[String] = Gen.alphaStr
-
-  implicit val variantGen: Gen[VariantT1] = for {
-    n <- Gen.choose(1, 5)
-    flds <- Gen.listOfN(n, Gen.zip(tagGen, singleTypeGen))
-  } yield VariantT1(RowT1(SortedMap(flds: _*), None))
-
-  def variantGenWithMandatoryEntry(mandatoryEntry: (String, TlaType1)): Gen[VariantT1] = variantGen.map { variantT =>
-    VariantT1(RowT1(variantT.row.fieldTypes + mandatoryEntry, None))
-  }
-
-  def variantGenWithMandatoryField(mandatoryField: String): Gen[VariantT1] =
-    singleTypeGen.flatMap { fldT =>
-      variantGenWithMandatoryEntry(mandatoryField -> fldT)
-    }
 
   test("variant") {
     type T = (String, TBuilderInstruction, VariantT1)
     type TParam = (String, TlaType1, VariantT1)
-
-    implicit val gen: Gen[TParam] = for {
-      tagName <- tagGen
-      valT <- singleTypeGen
-      variantT <- variantGenWithMandatoryEntry(tagName -> valT)
-    } yield (tagName, valT, variantT)
 
     // Variant(tagName, val) : varT
     def mkWellTyped(tparam: TParam): T = {
@@ -74,7 +50,7 @@ class TestVariantBuilder extends BuilderTest {
             mkIllTyped,
             resultIsExpected,
         )
-    )
+    )(Generators.tagValVariantGen)
 
     // throws on malformed variant type
     assertThrows[IllegalArgumentException] {
@@ -92,11 +68,6 @@ class TestVariantBuilder extends BuilderTest {
   test("variantFilter") {
     type T = (String, TBuilderInstruction)
     type TParam = (String, VariantT1)
-
-    implicit val gen: Gen[TParam] = for {
-      tagName <- tagGen
-      variantT <- variantGenWithMandatoryField(tagName)
-    } yield (tagName, variantT)
 
     // VariantFilter(tagName, set)
     def mkWellTyped(tparam: TParam): T = {
@@ -128,7 +99,7 @@ class TestVariantBuilder extends BuilderTest {
     def resultIsExpected = expectEqTyped[TParam, T](
         VariantOper.variantFilter,
         mkWellTyped,
-        { case (a, b) => Seq(builder.str(a), b) },
+        ToSeq.binary,
         { case (tagName, t) => SetT1(t.row.fieldTypes(tagName)) },
     )
 
@@ -139,7 +110,7 @@ class TestVariantBuilder extends BuilderTest {
             mkIllTyped,
             resultIsExpected,
         )
-    )
+    )(Generators.tagAndVariantGen)
 
   }
 
@@ -158,7 +129,7 @@ class TestVariantBuilder extends BuilderTest {
     def resultIsExpected = expectEqTyped[TParam, T](
         VariantOper.variantTag,
         mkWellTyped,
-        Seq(_),
+        ToSeq.unary,
         { _ => StrT1 },
     )
 
@@ -169,18 +140,12 @@ class TestVariantBuilder extends BuilderTest {
             mkIllTyped,
             resultIsExpected,
         )
-    )
+    )(Generators.variantGen)
   }
 
   test("variantGetOrElse") {
     type T = (String, TBuilderInstruction, TBuilderInstruction)
     type TParam = (String, TlaType1, VariantT1)
-
-    implicit val gen: Gen[TParam] = for {
-      tagName <- tagGen
-      valT <- singleTypeGen
-      variantT <- variantGenWithMandatoryEntry(tagName -> valT)
-    } yield (tagName, valT, variantT)
 
     // VariantGetOrElse(tagName, v, default)
     def mkWellTyped(tparam: TParam): T = {
@@ -216,7 +181,7 @@ class TestVariantBuilder extends BuilderTest {
     def resultIsExpected = expectEqTyped[TParam, T](
         VariantOper.variantGetOrElse,
         mkWellTyped,
-        { case (a, b, c) => Seq(builder.str(a), b, c) },
+        ToSeq.ternary,
         { case (_, t, _) => t },
     )
 
@@ -227,18 +192,13 @@ class TestVariantBuilder extends BuilderTest {
             mkIllTyped,
             resultIsExpected,
         )
-    )
+    )(Generators.tagValVariantGen)
 
   }
 
   test("variantGetUnsafe") {
     type T = (String, TBuilderInstruction)
     type TParam = (String, VariantT1)
-
-    implicit val gen: Gen[TParam] = for {
-      tagName <- tagGen
-      variantT <- variantGenWithMandatoryField(tagName)
-    } yield (tagName, variantT)
 
     // VariantGetUnsafe(tagName, v)
     def mkWellTyped(tparam: TParam): T = {
@@ -277,7 +237,7 @@ class TestVariantBuilder extends BuilderTest {
             mkIllTyped,
             resultIsExpected,
         )
-    )
+    )(Generators.tagAndVariantGen)
 
   }
